Moved workdir
