+++ ppqm/gaussian.py
@@ -6,10 +6,10 @@ from typing import Any, Callable, Dict, List, Optional, Union
 import numpy as np
 from tqdm import tqdm  # type: ignore[import]
 
-from ppqm import WorkDir, chembridge, constants, units
+from ppqm import chembridge, constants, units
 from ppqm.calculator import BaseCalculator
 from ppqm.chembridge import Mol
-from ppqm.utils import linesio, shell
+from ppqm.utils import WorkDir, linesio, shell
 
 G16_CMD = "g16"
 G16_FILENAME = "_tmp_g16_input.com"
